requirement.txt
flask
flask_cors
aiohttp
miservice_fork
unicodedata
pypinyin
mutagen